+++ PJATK-APBD-Cw7-s33986/DTOs/GetAllResponse.cs
@@ -3,7 +3,7 @@ namespace PJATK_APBD_Cw7_s33986.DTOs;
 public record GetAllResponse(
     int Id,
     string Name,
-    double Weight,
+    float Weight,
     int Warranty,
     DateTime CreatedAt,
     int Stock
