+++ PJATK-APBD-Cw7-s33986/DTOs/UpdatePcRequest.cs
@@ -7,7 +7,7 @@ public record UpdatePcRequest(
     [MaxLength(50)]
     string Name,
     [Required]
-    double Weight,
+    float Weight,
     [Required]
     int Warranty,
     [Required]
