+++ PJATK-APBD-Cw7-s33986/Models/PC.cs
@@ -4,7 +4,7 @@ public class PC
 {
     public int Id { get; set; }
     public string Name { get; set; } = string.Empty;
-    public double Weight { get; set; }
+    public float Weight { get; set; }
     public int Warranty { get; set; }
     public DateTime CreatedAt { get; set; }
     public int Stock { get; set; }
